Fixed Typos and Logic

+++ src/main/java/com/springboot/backend/repository/VideoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.springboot.backend.model.Video;
 
-public interface VideoRepository extends JpaRepository<Video, Long>{
+public interface VideoRepository extends JpaRepository<Video, Integer>{
 
 }
